init: add config option to specify I2P proxy
Add a new config option so that users can specify a dedicated SOCKS5
proxy for reaching I2P peers.

I2P is a closed network (almost). The I2P proxy cannot be used to reach
non-I2P peers and the other proxies (the generic one and the Tor one)
cannot be used to reach I2P peers. Thus the new option is unrelated to
the other options: does not default to `-proxy` and `-proxy` is not used
to reach I2P peers.

When the new option is set the I2P network will be set as reachable,
thus making the node relay I2P addresses to its peers (see
`RelayAddress()`).