rpc: fix "trusted" in WalletTxToJSON/TransactionDescriptionString
The TransactionDescriptionString() help doc used by RPCs
gettransaction, listtransactions, and listsinceblock
incorrectly stated that the "trusted" field was only
present if the transaction was trusted.

The "trusted" field was in fact only being returned by
WalletTxToJSON() when the transaction had zero confirmations.

It would make sense to return this boolean field in every case
and fix the help:

- update WalletTxToJSON() to always return the "trusted" field

- update TransactionDescriptionString() to the correct help doc for "trusted"